requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv
pydantic[email]